+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 2.x branch
 ## 2.15 branch
+### 2.15.2
+* added `beforeMarshal()` method do `PostsTable` class. Removed
+    `buildTagsForRequestData()`;
+* added `TagValidatorTrait` class. It provides some methods shared by the
+    validation classes;
+* fixed some little bugs for validation tags as string related to posts;
+* fixed bug, the tags of 3 characters were not accepted.
+
 ### 2.15.1
 * fixed some minor bugs for `Users` validation and added tests;
 * fixed `PostsTable::getRelated()` method and added tests;